+++ metro.config.js
@@ -4,4 +4,8 @@ const defaultConfig = getDefaultConfig(__dirname);
 
 module.exports = {
   ...defaultConfig,
+  resolver: {
+    ...defaultConfig.resolver,
+    assetExts: [...(defaultConfig.resolver?.assetExts ?? []), 'b64'],
+  },
 };